requirements.txt
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
oauthlib==3.1.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.5
publication==0.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.8.0
pylint==2.6.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
python-dateutil==2.8.1
python-dotenv==0.15.0
PyTweening==1.0.3
pytz==2021.1
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
robotframework==3.2.2
rpa==1.27.1
rsa==4.7.2
s3transfer==0.3.4
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
tagui==1.27.1
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.3.1
wcwidth==0.2.5
webdriver-manager==3.3.0
webencodings==0.5.1
wrapt==1.12.1
